% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_breaks.R
\name{cp_breaks}
\alias{cp_breaks}
\title{Calculate Map Breaks}
\usage{
cp_breaks(.data, var, newvar, classes, style)
}
\arguments{
\item{.data}{A tbl or sf object}

\item{var}{The existing variable to create classes of}

\item{newvar}{The new variable for storing the factor data}

\item{classes}{The number of breaks or classes to create}

\item{style}{The method for calculating classes, see classInt::classIntervals}
}
\description{
This function wraps the process I have been using to construction data classes for
choropleth maps using \code{ggplot2}.
}
