% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_examHelper.R
\name{cp_examHelper}
\alias{cp_examHelper}
\title{Multiple Choice Exam Key Creation}
\usage{
cp_examHelper(n)
}
\arguments{
\item{n}{number of questions in the exam}
}
\value{
A data frame containing the question numbers and correct answers
}
\description{
\code{cp_examHelper} creates an answer key for multiple choice
    exams that offer four possible answers in the form of the letters
    a, b, c, and d.
}
\examples{
key <- cp_examHelper(n = 40)
table(key$correct)

}
