% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_points.R
\name{cp_points}
\alias{cp_points}
\title{Points Conversion}
\usage{
cp_points(x, units = c("in", "cm", "mm"))
}
\arguments{
\item{x}{Value in points to be converted}

\item{units}{Size in units ("in", "cm", or "mm")}
}
\description{
This function converts points to other units of measurement for use with \code{ggsave()}.
}
\examples{
cp_points(940, units = "mm")

}
