% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_probt.R
\name{cp_probt}
\alias{cp_probt}
\title{p-value for t-test}
\usage{
cp_probt(t, df)
}
\arguments{
\item{t}{t value calculated from the given test}

\item{df}{Degrees of freedom}
}
\description{
This function returns the p-value for a t-test.
}
\examples{
cp_probt(3.6308, df = 72)

}
