% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_texTable.R
\name{cp_texTable}
\alias{cp_texTable}
\title{LaTeX Formatted Frequency Tables}
\usage{
cp_texTable(dataFrame, variable, title = "", type = c("doc", "table", "row"),
    sortByCount = FALSE, round = 2, missRow = TRUE, totalRow = TRUE)
}
\arguments{
\item{dataFrame}{Tibble or data frame object}

\item{variable}{Character, integer, double, or logical atomic vector}

\item{title}{Character string with the table's title}

\item{type}{Type of output - either full LaTex document ("doc"), table only ("table"), or rows only ("row")}

\item{sortByCount}{Logical scalar that determines if output will be sorted by count or name}

\item{round}{Number of decimal places displayed}

\item{missRow}{Logical scalar that determines if the NA row is displayed}

\item{totalRow}{Logical scalar that determines if the output will have a summary row appended}
}
\value{
a LaTeX formatted text block containing the counts and percentages of each value from the provided data
}
\description{
\code{cp_texTable} returns a LaTeX formatted frequency table with counts and percentages of values
  from a character, integer, double, or logical atomic vector
}
