% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTable.R
\name{freqTable}
\alias{freqTable}
\title{freqTable}
\source{
\href{https://cran.r-project.org/web/packages/frequencies/index.html}{\code{frequencies} package} (\href{https://cran.r-project.org/web/licenses/GPL-3}{released under GPL-3 license})
}
\usage{
freqTable(data_vector, sort_by_count = FALSE, round = 1, miss_row = TRUE,
  total_row = TRUE)
}
\arguments{
\item{data_vector}{an atomic vector of type character, integer, double, or logical}

\item{sort_by_count}{boolean value that determines if output will be sorted by count or name}

\item{round}{an integer value that determines the number of decimal places displayed}

\item{miss_row}{a boolean value that determines if the NA row is displayed}

\item{total_row}{a boolean value that determines if the output will have a summary row appended}
}
\value{
a data_frame containing the counts and percentages of each value from the provided data
}
\description{
\code{freqTable} returns a frequency table with counts and percentages of values
  from an atomic vector of type character, integer, double, or logical
}
