% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSave.R
\name{plotSave}
\alias{plotSave}
\title{Saving Plots at Prespecified Size and Resolution}
\usage{
plotSave(filename, plot, preset = c("sm", "med", "lg"))
}
\arguments{
\item{filename}{File name to create on disk; must be quoted}

\item{plot}{Plot to save, defaults to last plot displayed if no plot included.}

\item{preset}{Preset size ("sm", "med", "lg")}
}
\value{
Saves a file to the disk in the pre-determined size. \code{preset = "sm"}
}
\description{
This function is a wrapper around \code{ggsave()} that uses pre-specified values which
correspond to the possible graphics sizes in my lecture and presentation slides.
}
\examples{
\dontrun{
p <- ggplot(mpg, aes(hwy)) + geom_histogram()

plotSave("hwy.png", preset = "sm")
plotSave("hwy.png", p, preset = "med")
plotSave("hwy.png", p, preset = "lg")
}

}
