% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probt.R
\name{probt}
\alias{probt}
\title{p-value for t-test}
\usage{
probt(t, df)
}
\arguments{
\item{t}{t value calculated from the given test}

\item{df}{Degrees of freedom}
}
\description{
This function returns the p-value for a t-test.
}
\examples{
probt(3.6308, dof = 72)

}
