% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_sequoiaTheme.R
\name{cp_sequoiaTheme}
\alias{cp_sequoiaTheme}
\title{ggplot2 Theme for Sequoia Keynote Template}
\source{
\href{https://cran.r-project.org/web/packages/ggthemes/index.html}{\code{ggthemes} package} (\href{https://cran.r-project.org/web/licenses/GPL-2}{source code released under GPL-2 license})
}
\usage{
cp_sequoiaTheme(base_size = 28, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
The \code{cp_sequoiaTheme} is built on top of the \code{ggthemes}
package. It makes specific design changes to the \code{theme_fivethirtyeight} so
that the theme blends in seemlessly with the Alt slides in my Sequoia template
for Apple's Keynote. Among the changes completed are matching backgrounds, an
altered legend position and direction, a much larger default font size.
}
\examples{
library("ggplot2")
p1 <- ggplot(data = mpg) +
      geom_point(mapping = aes(x = hwy, y = displ, color = as.factor(cyl))) +
      cp_sequoiaTheme()

p2 <- ggplot(data = mpg) +
      geom_point(mapping = aes(x = hwy, y = displ, color = as.factor(cyl)),
      size = 4, position = "jitter") +
      scale_x_continuous(limits = c(10, 50)) +
      cp_sequoiaTheme() +
      labs(color = "Cylinder", x = "Highway Fuel Efficency", y = "Engine Displacement")

}
