% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_freqTable.R
\name{cp_tidyTable}
\alias{cp_tidyTable}
\title{Tidy Frequency Tables}
\source{
\href{https://cran.r-project.org/web/packages/frequencies/index.html}{\code{frequencies} package} (\href{https://cran.r-project.org/web/licenses/GPL-3}{source code released under GPL-3 license})
}
\usage{
cp_tidyTable(dataVector, sortByCount = FALSE, round = 1, missRow = TRUE, totalRow = TRUE)
}
\arguments{
\item{dataVector}{an atomic vector of type character, integer, double, or logical}

\item{sortByCount}{boolean value that determines if output will be sorted by count or name}

\item{round}{an integer value that determines the number of decimal places displayed}

\item{missRow}{a boolean value that determines if the NA row is displayed}

\item{totalRow}{boolean value that determines if the output will have a summary row appended}
}
\value{
a tibble containing the counts and percentages of each value from the provided data
}
\description{
\code{cp_tidyTable} returns a frequency table with counts and percentages of values
  from an atomic vector of type character, integer, double, or logical
}
\examples{
tableData <- data.frame(
   id = c(1, 2, 3, 4, 5, 6),
   nhoodStr = c("Patch", "Bevo Mill", "Bevo Mill", "Lindenwood Park", "Carondelet" , "Shaw"),
   visit = c(TRUE, FALSE, FALSE, TRUE, TRUE, NA),
   type = as.factor(c("A", "B", "B", "A", "B", "B")),
   stringsAsFactors = FALSE
   )

cp_tidyTable(tableData$nhoodStr)
cp_tidyTable(tableData$nhoodStr, round = 2)
cp_tidyTable(tableData$visit, round = 2, missRow = TRUE)
cp_tidyTable(tableData$visit, round = 2, missRow = FALSE)
cp_tidyTable(tableData$type, round = 3, sortByCount = FALSE)
cp_tidyTable(tableData$type, round = 3, sortByCount = TRUE)

}
